grade_calculate/requirements.txt
openai
requests
json
python-dotenv